python/requirements.txt
boto3
oss2
pyyaml